backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing & Computer Vision
pillow>=10.3.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0

# Machine Learning - YOLOv8
ultralytics>=8.2.0
torch>=2.3.0
torchvision>=0.18.0

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication & Security
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# Configuration
python-dotenv==1.0.0
pydantic>=2.9.2

# Advanced Training & Evaluation (for model development)
pycocotools>=2.0.7          # COCO dataset format handling
shapely>=2.0.0              # Spatial geometry (in_bin detection)
albumentations>=1.3.0       # Advanced image augmentations
scikit-learn>=1.3.0         # Precision-recall curves, evaluation
matplotlib>=3.7.0           # Visualization and plots
tqdm>=4.66.0                # Progress bars